squash: merged `AbstractProcessEngineAware` into `AbstractRestProcessEngineAware`
