--- engine-rest/engine-rest/src/main/java/org/camunda/bpm/engine/rest/spi/impl/AbstractProcessEngineAware.java
@@ -1,42 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information regarding copyright
- * ownership. Camunda licenses this file to you under the Apache License,
- * Version 2.0; you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.camunda.bpm.engine.rest.spi.impl;
-
-import javax.ws.rs.core.Response.Status;
-import org.camunda.bpm.engine.ProcessEngine;
-import org.camunda.bpm.engine.rest.exception.InvalidRequestException;
-import org.camunda.bpm.engine.rest.util.EngineUtil;
-
-public abstract class AbstractProcessEngineAware {
-
-  private ProcessEngine processEngine;
-
-  public AbstractProcessEngineAware() {
-    processEngine = EngineUtil.lookupProcessEngine(null);
-  }
-
-  public AbstractProcessEngineAware(String engineName) {
-    processEngine = EngineUtil.lookupProcessEngine(engineName);
-  }
-
-  protected ProcessEngine getProcessEngine() {
-    if (processEngine == null) {
-      throw new InvalidRequestException(Status.BAD_REQUEST, "No process engine available");
-    }
-    return processEngine;
-  }
-}
